Realized that there is not really common code in the channel examples so made herd_to_herd standalone too. Also updated docs
